+++ CHANGELOG.md
@@ -13,6 +13,18 @@
 - Renamed `PlaybackCompletionEvent` to `PlaybackCompletion`
 - Switched to `recursive_spawn` for `sample_effects`, permitting an
   arbitrary number of effects to be spawned at once
+- Updated Firewheel backend management.
+
+### Migration guide
+
+Firewheel's `cpal` backend now waits until an API has been selected before
+providing its simple device configuration. As a result, device entities will
+not be available until after `SeedlingStartupSystems::StreamInitialization`
+in `PostUpdate`, and the stream is automatically started with the default
+output.
+
+Any code that relies on device entities in the `Startup` schedule will need
+to be moved to `PostStartup`.
 
 # 0.6.1